fix(apidocs): move type AsyncApiResolver and config to AsyncApiDefinition component
