+++ plugins/api-docs/src/components/AsyncApiDefinitionWidget/AsyncApiDefinitionWidget.tsx
@@ -16,6 +16,7 @@
 
 import { Progress } from '@backstage/core-components';
 import React, { Suspense } from 'react';
+import { AsyncApiResolver } from './AsyncApiDefinition';
 
 // The asyncapi component and related CSS has a significant size, only load it
 // if the element is actually used.
@@ -24,14 +25,6 @@ const LazyAsyncApiDefinition = React.lazy(() =>
     default: m.AsyncApiDefinition,
   })),
 );
-
-/** @public */
-export type AsyncApiResolver = {
-  schema: string;
-  order: number;
-  canRead: boolean;
-  read(uri: any): Promise<string>;
-};
 
 /** @public */
 export type AsyncApiDefinitionWidgetProps = {
