+++ plugins/api-docs/src/components/AsyncApiDefinitionWidget/index.ts
@@ -15,7 +15,5 @@
  */
 
 export { AsyncApiDefinitionWidget } from './AsyncApiDefinitionWidget';
-export type {
-  AsyncApiDefinitionWidgetProps,
-  AsyncApiResolver,
-} from './AsyncApiDefinitionWidget';
+export type { AsyncApiDefinitionWidgetProps } from './AsyncApiDefinitionWidget';
+export type { AsyncApiResolver } from './AsyncApiDefinition';
